Force redactor initialization to wait DOM load
In some browsers, "$.redactor()" is not detected as a function because
the plugin script code still hasn't finished loading, resulting in a
fatal error.

Wrapping the $.redactor() function in $.ready() forces to wait until the
DOM is ready, thus making sure the function is ready to be used (and the
involved DOM is ready to be manipulated).